changed flag to float

+++ app.py
@@ -14,9 +14,9 @@ parser.add_argument('--epochs', type=int, default=20)
 
 parser.add_argument('--batch_size', type=int, default=128)
 
-parser.add_argument('--num_epochs_per_decay', type=int, default=50.)
+parser.add_argument('--num_epochs_per_decay', type=float, default=50.)
 
-parser.add_argument('--lr_decay_factor', type=int, default=0.1)
+parser.add_argument('--lr_decay_factor', type=float, default=0.1)
 
 parser.add_argument('--dev', type=bool, default=False)
